Configure Git: add alias for more accurate time
Added an alias that I will use in place of 'git commit' which should provide a more accurate author date for all of my commits.

The script looks for the latest modified date for each of the files about to be committed, and uses that as the author date instead of making it the same as the commit date.